requirements.txt
django
djangorestframework
djangoframework-simplejwt